Enhancement: Async Validation now works (the Upload Service now calls the new Validation Service asynchronously)
