+++ sodalite-server/src/main/java/org/icgc/dcc/sodalite/server/service/UploadService.java
@@ -5,19 +5,11 @@ import static org.springframework.http.ResponseEntity.ok;
 
 import org.icgc.dcc.sodalite.server.model.Upload;
 import org.icgc.dcc.sodalite.server.repository.UploadRepository;
-import org.icgc.dcc.sodalite.server.validation.SchemaValidator;
 import org.skife.jdbi.v2.exceptions.UnableToExecuteStatementException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.fasterxml.jackson.module.paramnames.ParameterNamesModule;
 
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
@@ -32,12 +24,7 @@ public class UploadService {
   @Autowired
   IdService id;
   @Autowired
-  private SchemaValidator validator;
-
-  protected static final ObjectMapper mapper = new ObjectMapper().registerModule(new ParameterNamesModule())
-      .registerModule(new Jdk8Module())
-      .registerModule(new JavaTimeModule());
-
+  ValidationService validator;
   @Autowired
   private final UploadRepository uploadRepository;
 
@@ -55,7 +42,7 @@ public class UploadService {
       throw new RepositoryException(jdbie.getCause());
     }
 
-    validate(schemaName, uploadId, payload); // Async operation.
+    validator.validate(schemaName, uploadId, payload); // Async operation.
 
     return ok(uploadId);
   }
@@ -77,37 +64,8 @@ public class UploadService {
     return ok("Successfully published " + uploadId);
   }
 
-  @Async
-  // TODO: Move this into another class so that Spring can call it asynchronously.
-  private void validate(String schemaId, String uploadId, String payload) {
-    try {
-      val jsonNode = mapper.reader().readTree(payload);
-      val response = validator.validate(schemaId, jsonNode);
-
-      if (response.isValid()) {
-        updateAsValid(uploadId);
-      } else {
-        updateAsInvalid(uploadId, response.getValidationErrors());
-      }
-    } catch (JsonProcessingException jpe) {
-      log.error(jpe.getMessage());
-      updateAsInvalid(uploadId, format("Invalid JSON document submitted: %s", jpe.getMessage()));
-    } catch (Exception e) {
-      log.error(e.getMessage());
-      updateAsInvalid(uploadId, format("Unknown processing problem: %s", e.getMessage()));
-    }
-  }
-
   private void save(@NonNull String studyId, @NonNull String uploadId, @NonNull String jsonPayload) {
     uploadRepository.create(uploadId, studyId, Upload.CREATED, jsonPayload);
-  }
-
-  private void updateAsValid(@NonNull String uploadId) {
-    uploadRepository.update(uploadId, Upload.VALIDATED, "");
-  }
-
-  private void updateAsInvalid(@NonNull String uploadId, @NonNull String errorMessages) {
-    uploadRepository.update(uploadId, Upload.VALIDATION_ERROR, errorMessages);
   }
 
   private void updateAsPublished(@NonNull String uploadId) {
